[WIP] Queue events when offline
It looks like Socket.io already does some queing automatically.
I noticed that if I disconnected using the
hacky mechanism commented out in the server index.js,
that all I had to do was call socket.connect() again
and it would complete sending the `propose`.
However, I'm not sure if that works with all kinds of disconnects.

At the end of the day, this PR mostly creates hooks for being able to display whether we're online or offline, as well as a hook for reconnecting. This all needs to get wired up into Elm.

It also starts attempting to track the client's event id instead of having all
events have a clientId of zero. However, given that Socket.io already does its own queueing,
it looks like this may not have been necessary.

Lastly, when reconnecting, we're not doing anything yet on client or server to
filter catchup events to only ones we don't have.