Reorganize repo to allow for multiple examples
This works okay. I had a hard time figuring out a good way to share
the server / websocket code, which is likely to be identical
between examples. I was trying to do it with a symlink but
maybe it'd be better to do using a require. For now it's
just copied.